🚚 Move tuple arbitraries into /arbitrary
