+++ src/check/arbitrary/FloatingPointArbitrary.ts
@@ -2,7 +2,7 @@ import { Arbitrary } from './definition/Arbitrary';
 import { doubleNext, DoubleNextConstraints } from './DoubleNextArbitrary';
 import { floatNext, FloatNextConstraints } from './FloatNextArbitrary';
 import { integer } from '../../arbitrary/integer';
-import { tuple } from './TupleArbitrary';
+import { tuple } from '../../arbitrary/tuple';
 
 /** @internal */
 function next(n: number): Arbitrary<number> {
