+++ src/check/arbitrary/FunctionArbitrary.ts
@@ -4,7 +4,7 @@ import { cloneMethod, hasCloneMethod } from '../symbols';
 import { array } from '../../arbitrary/array';
 import { Arbitrary } from './definition/Arbitrary';
 import { integer } from '../../arbitrary/integer';
-import { tuple } from './TupleArbitrary';
+import { tuple } from '../../arbitrary/tuple';
 import { escapeForMultilineComments } from './helpers/TextEscaper';
 
 /**
