+++ src/check/arbitrary/WebArbitrary.ts
@@ -8,7 +8,7 @@ import { ipV4, ipV4Extended, ipV6 } from './IpArbitrary';
 import { oneof } from '../../arbitrary/oneof';
 import { option } from '../../arbitrary/option';
 import { stringOf } from './StringArbitrary';
-import { tuple } from './TupleArbitrary';
+import { tuple } from '../../arbitrary/tuple';
 import { Arbitrary } from './definition/Arbitrary';
 
 /**
