+++ src/fast-check-default.ts
@@ -77,7 +77,8 @@ import {
   unicodeString,
 } from './check/arbitrary/StringArbitrary';
 import { shuffledSubarray, subarray, SubarrayConstraints } from './check/arbitrary/SubarrayArbitrary';
-import { genericTuple, tuple } from './check/arbitrary/TupleArbitrary';
+import { genericTuple } from './arbitrary/genericTuple';
+import { tuple } from './arbitrary/tuple';
 import { uuid, uuidV } from './check/arbitrary/UuidArbitrary';
 import {
   webAuthority,
